+++ presentation/src/androidTest/java/com/delivery/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.dsm_delivery
+package com.delivery
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
