+++ presentation/src/main/java/com/delivery/presentation/base/BaseActivity.kt
@@ -1,0 +1,29 @@
+package com.delivery.presentation.base
+
+import android.os.Bundle
+import android.widget.Toast
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+
+abstract class BaseActivity<B : ViewDataBinding>(
+    @LayoutRes private val layoutId: Int,
+) : AppCompatActivity() {
+
+    protected lateinit var binding: B
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = DataBindingUtil.setContentView(this,layoutId)
+        binding.lifecycleOwner = this
+
+        initView()
+    }
+    fun showShortToast(msg: String){
+        Toast.makeText(this,msg,Toast.LENGTH_SHORT).show()
+    }
+
+    abstract fun initView()
+}
