+++ presentation/src/main/java/com/delivery/presentation/feature/login/LoginActivity.kt
@@ -1,0 +1,17 @@
+package com.delivery.presentation.feature.login
+
+
+import com.delivery.presentation.base.BaseActivity
+import dsm.delivery.presentation.R
+import dsm.delivery.presentation.databinding.ActivityLoginBinding
+
+
+class LoginActivity : BaseActivity<ActivityLoginBinding>(
+    R.layout.activity_login
+){
+    override fun initView() {
+        //TODO("Not yet implemented")
+
+    }
+
+}
