+++ presentation/src/main/java/com/delivery/presentation/feature/main/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.delivery.presentation.feature.main
+
+
+import com.delivery.presentation.base.BaseActivity
+import dsm.delivery.presentation.R
+import dsm.delivery.presentation.databinding.ActivityMainBinding
+
+class MainActivity : BaseActivity<ActivityMainBinding>(
+    R.layout.activity_main
+){
+    override fun initView() {
+        //TODO 일단 나중...
+
+
+
+    }
+
+}
