--- presentation/src/main/java/presentation/base/BaseActivity.kt
@@ -1,4 +1,0 @@
-package presentation.base
-
-class BaseActivity {
-}
