--- presentation/src/main/java/presentation/feature/login/LoginActivity.kt
@@ -1,13 +1,0 @@
-package presentation.feature.login
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import dsm.delivery.presentation.R
-
-
-class LoginActivity : AppCompatActivity(){
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-    }
-}
