--- presentation/src/main/java/presentation/feature/main/MainActivity.kt
@@ -1,14 +1,0 @@
-package presentation.feature.main
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import dsm.delivery.presentation.R
-
-
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
