+++ presentation/src/main/res/layout/activity_mypage.xml
@@ -89,7 +89,7 @@
 
 
         <androidx.cardview.widget.CardView
-            android:layout_width="350dp"
+            android:layout_width="360dp"
             android:layout_height="150dp"
             android:layout_marginTop="15dp"
             app:cardCornerRadius="10dp"
